// src/shared/api/generated/models/reportingMetricCellDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportingMetricCellDtoState } from "./reportingMetricCellDtoState";
import type { ReportingMetricCellDtoSuppressionReason } from "./reportingMetricCellDtoSuppressionReason";

export interface ReportingMetricCellDto {
  code: string;
  sampleSize?: number;
  state: ReportingMetricCellDtoState;
  suppressionReason?: ReportingMetricCellDtoSuppressionReason;
  value?: string;
}