// src/shared/api/generated/models/reportingQueryResultResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportingResourceReceiptDto } from "./reportingResourceReceiptDto";
import type { ReportingResultDocumentDto } from "./reportingResultDocumentDto";
import type { ReportingQueryResultResponseDtoStatus } from "./reportingQueryResultResponseDtoStatus";

export interface ReportingQueryResultResponseDto {
  consumerId?: string;
  queryHash: string;
  receipt?: ReportingResourceReceiptDto;
  result?: ReportingResultDocumentDto;
  resultId?: string;
  retryAfterMs?: number;
  runId: string;
  status: ReportingQueryResultResponseDtoStatus;
}