// src/shared/api/generated/models/reportingResultRowDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportingResultRowDtoDimensions } from "./reportingResultRowDtoDimensions";
import type { ReportingMetricCellDto } from "./reportingMetricCellDto";

export interface ReportingResultRowDto {
  currency?: string;
  day?: string;
  dimensions?: ReportingResultRowDtoDimensions;
  metrics: ReportingMetricCellDto[];
}